[Gold IV] Title: 타임머신, Time: 8 ms, Memory: 2288 KB -BaekjoonHub

+++ 백준/Gold/11657. 타임머신/타임머신.cc
@@ -4,7 +4,7 @@ typedef long long ll;
 typedef tuple<int, int, ll> tup;
 
 int n, m;
-const int INF = (int)1e18;
+const ll INF = (ll)1e18;
 vector<tup> adj;
 vector<ll> dist;
